Reduce keystone cache expiration time
This commit removes the keystone_cache_expiration_time variable,
allowing deployers to override using standard config overrides.

Additionally, we /temporarily/ disable catalog caching in
keystone.conf.j2 to unblock our master gate.

Note that this change is necessary as we have been experiencing
fairly consistent gate failures in master where the tempest role is
unable to upload an image to glance.  This causes the tempest run to
fail since the expected image is not available.  This upstream keystone
review [1] introduced a change to add catalog caching, and what we are
actually seeing is glance failing to upload the image (or create the
necessary directory in swift) as the object-store endpoint is not in
the catalog which it gets back from keystone.  When this happens we are
presumably hitting a stale cache entry which will not expire for 5400
seconds (the old value of keystone_cache_expiration_time).

For some additional information (as provided by dolphm), it looks like
endpoint updates invalidate portions of the dogpile cache but not all.
There is a fix in flight [2] to address this.

[1] https://review.openstack.org/#/c/215212/
[2] https://review.openstack.org/#/c/271536/

UpgradeImpact
Related-bug: #1537617

Change-Id: I3b2a833a3e96a3b5deac76052eed480ddcb6175d