Enable functional testing
This commit uncomments the disablement of the functional test in
tox.ini and also overrides keystone_git_install_branch by pointing to
the SHA currently in openstack-ansible.  This is necessary as master of
keystone will not build using this role due to changes in upstream
keystone.  We can tackle that issue once we have more roles performing
functional testing.

Change-Id: Ia2323fd48c143798ea5fd2815c9b83586d16b0e6
Partial-Bug: #1553957