Use upper-constraints file in developer mode
Currently, keystone fails to install in developer mode due to us not
specifying upper constraints.  This is due to an upcapped version of
pysaml2 being installed which is causing problems.  This commit first
clones the requirements repo and then uses the upper constraints file
when installing packages.

Note that we only update pip_install_options when in developer mode if
the constraints aren't already set.  This is because in developer mode
you could have a container with multiple services running on it and the
pip_install_options fact would get updated when already set, resulting
in pip failing to install packages.

Change-Id: I4f21c20a1f7fa1fc4277cdffc98fc9b2017f11d1