Fix how constraints are added to pip_install_options
This commit changes the pip_install_options fact name in
keystone_install.yml to pip_install_options_fact.  This allows us to
maintain the existing pip_install_options variable without overwriting
it with options when in developer mode, which ultimately means we can
have multiple services running in a container use a combination of
*_developer_mode: true and *_developer_mode: false.  At the moment,
if a service writes pip_install_options fact with the constraints
options, those options will persist to other services running in the
container even if *_developer_mode: false.

If this change is accepted, the same will need to be done to all the
IRRs.

Change-Id: Ic76074da419685148e27f2f1295ddd2d82bb03f9