Use ansible facts for distributing SSL certs/keys
The os_keystone role previously relied on a memcached deployment to transfer
SSL certificates and keys to all keystone nodes. Many of the openstack-ansible
repositories have refactored this behavior out in place of registering the
certificates and keys as ansible facts and using ansible's copy module to
transfer them to each node in the deployment.

This breaks the dependency of requiring memcached in order to deploy keystone
with SSL.

Change-Id: I8db39a2a4a54aa9814c1b05988f05bfcae94f222