Skip unavailable hosts when distributing keys
The 'Create authorized keys file from host vars' task assumes that every
host within the keystone_all group is available. If any host within that
group is offline, excluded through a limit, or has otherwise failed an
earlier task, it will lead to this task failing for every host.

Add a when clause checking for the existence of the keystone_pubkey
hostvar so that the task and remainder of the role can continue when a
keystone_all host is unavailable.

Change-Id: Ie68e537872fec82769c8fffcf5e144d695de59df