Update the keystone WSGI application locations
This change will update the file paths of the keystone WSGI
applications so that keystone will run properly and not just
produce errors in the logs. The virtualenvs were the primary reason
for this bug, and this solution allows for the deployment of
keystone without requiring virtualenvs provided that they override
keystone_bin.

Closes-bug: 1586161

Change-Id: Iedf4c2651d304d3034be64c781a6daf7ab0b400f
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>