Force a restart of all the apache nodes during upgrade
During the upgrades, the venv path will change and therefore the
apache configuration file will change too.

However we apply the restart of apache after the keystone_service_setup
for nodes [1:] (the first node gets restarted as first task of the
keystone_service_setup).

So during an upgrade, because apache is up, the configuration file has
changed but apache still serves the old code (because not restarted yet
on the nodes 1 and above) when the keystone_service_setup is applied.

The keystone module can then hit any node in the load balancer, which
could be a different version.

This commit fixes the issue by ensuring apache is restarted and
therefore runs the latest code.

Change-Id: Iac94a8fc337c2139d1876b9753e46815910a0ba0
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>