Add tempest to keystone role tests
The keystone functional tests are currently not using tempest, and are
quite lacking in functionality.

This PR adds tempest testing to the keystone role.
We use the tempest.api.identity tests but exclude the tests for
credentials and ec2 which we don't currently support.

TODO: (andymccr)
We need to add support for these (and remove the exclusion) by
performing a 'keystone-manage credential_setup --keystone-user keystone
--keystone-group keystone'

Change-Id: I9a7207e75040c304c53820795cb66ce9be00c350