Add credential_setup for keystone
We need to setup the appropriate directory for credential_setup and run
the keystone-manage credential_setup command.

We created the directory and the '[credential]' stanza in the
keystone.conf, which will ensure we can add additional settings using
config_template if any further are required.

We need to setup the autorotation cron job and distribution for
credential keys.

Additionally, we include all tempest tests now that we are
supporting this feature.

Change-Id: Ifd85ed1a64538ed037e4426cc50238d2b16d51e5