Simplify pip options/constraints mechanism
The current constraints generation for the
installation involves multiple tasks and multiple
variables.

Using multiple tasks extends the installation time
unnecessarily and the additional variables are
unnecessary.

This patch aims to simplify the mechanism and
hopes to speed it up a little.

The tests/os_keystone-overrides.yml file is removed
as it is unnecessary given the merge of
https://review.openstack.org/381479

Depends-On: I0e6d43a9f22add78f693e8836f1dd852be74d6c5
Change-Id: Ief9c7f124eb58b986192b2f958c971197d410799