CentOS: Only install Federation IDP/SP Packages when necessary
In https://review.openstack.org/309425 adjustments were
made which resulted in keystone_idp and keystone_sp to
always be defined. Unfortunately the CentOS support for
os_keystone merged after that and reviews did not pick
up the necessary changes.

This patch corrects it so that the repo and packages are
only installed if necessary.

Additionally, the Federation SP callback template was
being unnecessarily copied over. It will now only copy
over when it will be used.

Change-Id: I466f1391893d33be7c83caba70ec93e44dd1d482